+++ src/Feedy.Application/DependencyInjection.cs
@@ -1,6 +1,7 @@
 namespace Feedy.Application;
 
 using Feedy.Application.UseCases.GetRecipeFeed;
+using Feedy.Application.UseCases.LoginUser;
 using Feedy.Application.UseCases.RegisterUser;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -11,8 +12,9 @@ public static class DependencyInjection
         services.AddScoped<FeedRankingService>();
         services.AddScoped<GetRecipeFeedService>();
 
-        services.AddScoped<PasswordHasher>();
+        // IPasswordHasher and IJwtTokenProvider are registered in Feedy.Infrastructure
         services.AddScoped<RegisterUserService>();
+        services.AddScoped<LoginUserService>();
 
         return services;
     }